chore: move all __main__ to app level
